+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased] - YR
+### Changed
+- Fixed models/Task.php to check if string for task name
+
+### Added
+- Many more tests in tasktest.php
+
+## [Unreleased] - YR
 ### Moved
 - Entity.php from application/models to application/core